refactor(memory): extract RegionSizeSpec from BundleLayoutCalculator
Extract RegionSizeSpec to a standalone public record to improve modularity and clean up the bundle layout calculator class. Update all references and unit tests.
