+++ memory/spector-memory/src/main/java/com/spectrayan/spector/memory/kernel/bundle/BundleLayoutCalculator.java
@@ -24,17 +24,6 @@ public final class BundleLayoutCalculator {
     
     private BundleLayoutCalculator() {} // static utility
     
-    /** Specification for a single region's size requirements. */
-    public record RegionSizeSpec(
-        RegionId regionId,
-        long dataBytes,      // total data bytes (excluding region's own SMKM header)
-        int capacity,        // max records
-        int stride,          // record stride (from store's MemoryLayout)
-        int layoutId,        // store's MemoryLayout.layoutId()
-        int schemaVersion,   // store's MemoryLayout.schemaVersion()
-        boolean growable     // whether region can grow via relocate-to-tail
-    ) {}
-    
     /**
      * Computes a complete bundle layout from region specs.
      * Returns (BundleDirectory, totalFileSize).
